extract alexa configs
