+++ alexa-spring-plugin/src/main/kotlin/org/rewedigital/dialog/alexa/spring/IntentHandlerConfig.kt
@@ -1,4 +1,4 @@
-package org.rewedigital.dialog.springsample.alexa
+package org.rewedigital.dialog.alexa.spring
 
 import com.amazon.ask.dispatcher.request.handler.RequestHandler
 import org.rewedigital.dialog.spring.annotations.IntentHandler
