+++ settings.gradle
@@ -6,3 +6,4 @@ include 'spring-sample'
 include 'spring-plugin'
 include 'konversation-plugin'
 include 'alexa-plugin'
+include 'alexa-spring-plugin'
