+++ ssml-plugin/build.gradle
@@ -9,9 +9,10 @@ apply from: '../docu.gradle'
 apply plugin: 'com.novoda.bintray-release'
 
 dependencies {
+    implementation project(":core")
+    implementation project(":ssml")
+
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
-    implementation "org.rewedigital.voice:dialog:${rootProject.ext.versions.core}"
-    implementation "org.rewedigital.voice:ssml-builder:${rootProject.ext.versions.ssml}"
 }
 
 dokka {
